Addendum: ensure the server sees the sign in
next@6 had a subtle change that meant the server wasn't seeing the query
token, but also there was a logic bug in my own code that meant when the
user signed in for the first time, they wouldn't be signed in until a
page refresh (either server side or client side).

This change fixes that logic and ensures the code is always able to see
the token when it's being passed in for the first time.